split DeviceRepository into two classes for SRP
